Use single precision

+++ src/lib.rs
@@ -1,14 +1,14 @@
 use rayon::prelude::*;
 
 const MAX_ITER: i32 = 1000;
-const RADIUS_SQ: f64 = 4.0;
+const RADIUS_SQ: f32 = 4.0;
 
-const XMIN: f64 = -2.5;
-const XMAX: f64 = 1.0;
-const YMIN: f64 = -1.0;
-const YMAX: f64 = 1.0;
+const XMIN: f32 = -2.5;
+const XMAX: f32 = 1.0;
+const YMIN: f32 = -1.0;
+const YMAX: f32 = 1.0;
 
-pub fn lsm(cr: f64, ci: f64) -> i32 {
+pub fn lsm(cr: f32, ci: f32) -> i32 {
     let mut zr = 0.0;
     let mut zi = 0.0;
     let mut zr2 = 0.0;
@@ -31,8 +31,8 @@ pub fn render_mandelbrot(buffer: &mut [u32], width: usize, height: usize) {
         .for_each(|(y, rows)| {
             rows.iter_mut().enumerate().for_each(|(x, pixel)| {
                 let (cr, ci) = (
-                    (x as f64 / width as f64) * (XMAX - XMIN) + XMIN,
-                    (y as f64 / height as f64) * (YMAX - YMIN) + YMIN,
+                    (x as f32 / width as f32) * (XMAX - XMIN) + XMIN,
+                    (y as f32 / height as f32) * (YMAX - YMIN) + YMIN,
                 );
                 let iterations = lsm(cr, ci);
